Update.
Switch to RubyGems style project layout. Structure, style, document and refine
the code overall.

Provide the same USER, HOME, SHELL and PWD in the container through the use of
a base image.

Set name and hostname for the container launched as packagename-pid. Setting
hostname is helpful as most users display hostname in their shell prompt.
Setting name is helpful to identify the container in `docker ps` listing.

Remove container after execution, so as to not pile up dead containers. This
one might be debatable.

Basic command line interface.

Signed-off-by: Saurabh Kumar <saurabhgeek92@gmail.com>
Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>