Don't set $shell as login shell in baseimage's _init script.
Thus if user changes his login shell on the host, it will be reflected in the
container too.

---

$shell in container is the same as $SHELL on the host. And $SHELL is always a
pointer to the user's login shell, not the current shell. So switch only ever
launches user's login shell in the container.

Behaviour of switch is much like opening a new tab / terminal window: launch
login shell and set PWD from the focussed tab or window.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>