Change approach.
Instead of building new Dockerfiles from our baseimage, switch-ify any given
Dockerfile. This requires us to rely on heuristics to detect OS of the image
given to us. We make it work for Debian, Ubuntu and CentOS for now, and will
improve our approach and incorporate other baseimages over time.

We also ensure that the user we create in the docker image has the same UID and
GID as on the host. It must be noted that on OX X boot2docker image is the real
host, and thus corresponding UID & GID are used.

close #12

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>