deb: Ensure always system Ruby will be used.
Rewrite shebang line to `#!/usr/bin/ruby` instead of `#!/usr/bin/env ruby`.
This ensures the package always will always use system Ruby regardless of
the Ruby loaded in the env by rvm, rbenv, chruby, etc. This is the default
behaviour in Jessie and Vivid, so the fix is meant only for Trusty.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>